+++ docs/changes.md
@@ -10,6 +10,7 @@ title: Memory Change Log
 * Add `Orchestra\Memory\Abstractable\Container`.
 * Predefined package path to avoid additional overhead to guest package path.
 * Rename command to `php artisan memory:migrate`.
+* Implement caching for both eloquent and fluent driver.
 
 ## Version 2.0 {#v2-0}